requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dlib==19.24.2
Flask==3.0.2
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
python-dotenv==1.0.1
Werkzeug==3.0.1
